Move autofillConstants to autofill folder
